+++ NEWS.md
@@ -12,6 +12,7 @@ output:
 ## CAVAanalytics 1.1.3
 - Added function model_biases to look at the delta between observation and model simulations
 - Customized cli messages and improved clarity
+- Improved visualization of spatial-temporal plots using ggridges
 
 ## CAVAanalytics 1.1.2
 - Changed appearance of temporal trends when spatial.aggr=T